Generation of manifests to TypeScript
When using the Umbrella script it is now possible to generate it toTypeScript using the --typescript option. Also the script checks if the checksum of the file is the same. If the file is the same it will skip writing to disk. Added switch -d for doing a dry run